wip: select star from databases where theme is elephant (#7)
* feat: add docker-compose.yml

I'm not the biggest fan of docker-compose, but phew, it sure beats
manually aligning everyone's dev setup.

* feat: rework middleware

Middleware can be provided arguments now.

For the most part these'll always be empty, but
things like Postgres may want to allow overrides
in the future.

* dep: add deps on cls-hooked, ormnomnom, pg

* lint: so much lint

* lint: continued futzing with prettier/eslint.

* feat: transaction middleware

* fix: it sure helps to include the file you just wrote

* WIP! rearrange all of the furniture in the room

put the couch on the ceiling. chairs on the walls! carefully make the rug
levitate in the center of the room. set it just so, gently spinning.