Changes to when we log.
We want to log after the request lifecycle is complete, in some kind of
post hook. Middlewares don't really give this to us.

For now, we log at the end of the request handler, which is when we have
all the information we need. If we need more than one thing to run after
the request has been sent, we'll revisit this.