And finally, a router/muxer/whatever.
We're using a tiny wrapper around find-my-way that I'm not even
sure we need. But it's a teensy bit of sugar.

The INTERESTING THING TO NOTE is that our context object ends up being
treated by find-my-way as the `this` bound in the route handlers. This
is not so bad, I think. Don't name any function arguments in your
handler; use only `this`, return a Response: boom.