Convulsions in how we route and wrap requests (#10)
* Request contexts: wrap raw req/res in an object.

This object is now used to hold request lifecycle data instead
of the request object. Callers may consult the res object, but
probably should not ever call methods on it.

* Changes to when we log.

We want to log after the request lifecycle is complete, in some kind of
post hook. Middlewares don't really give this to us.

For now, we log at the end of the request handler, which is when we have
all the information we need. If we need more than one thing to run after
the request has been sent, we'll revisit this.

* And finally, a router/muxer/whatever.

We're using a tiny wrapper around find-my-way that I'm not even
sure we need. But it's a teensy bit of sugar.

The INTERESTING THING TO NOTE is that our context object ends up being
treated by find-my-way as the `this` bound in the route handlers. This
is not so bad, I think. Don't name any function arguments in your
handler; use only `this`, return a Response: boom.

* I no longer know what the hell is going on with my semicolons.

* Awaiting send().