The simplest possible whitelist. (#19)
Plus, a whole lot of improvements on the read-through cache for
VCpm packages. We now handle errors thrown by pacote, and make some
kind of reasonable response. I'm not sure it's the response the VC
cli wants, but I can deal with that later.

The whitelist is a newline-delimited list of package names, pointed
to by the env var WHITELIST. That's it.