Fix issue #34: Implement CSRF checking
Added the `csrf` package for generating random tokens. Wrapped the
tokens handlers in a function that does a couple of things. For read
http verbs, it stores a secret in the session and makes a matching token
available on the context object for use in forms. Made our forms use it.

For mutating http verbs, the wrapper checks the token against the stored
secret, and redirects to login if it's not a match.