fix issue #32: implement whoami (#35)
* issue #32: implement whoami

This isn't all that important from a legacy API, but is a building block
for all future authenticated actions in our own API. We now have a
middleware that decorates the context object with a user if bearer auth
is present with a valid token.

Also, ran lint.

* Indulge my need for consistent sugar.

Make all the response conveniences have the same basic signature,
so the pattern is predictable. You can override status & add headers
for each of them now.

Added a new response convenience for auth-needed.

* Improve our bearer auth middlware.

Also, only respond with *active* tokens and *active* users.