Groundwork for package & user name validations.
Wrap up the legacy package name validity check to hide the parts of that
API we don't care about. Sprinkle a little joi here & there.

More importantly, implement a hook for defining entropic's package name rules.

I have commenced the bikeshedding with an opinion: a-z, 0-0, and -.
Controversially, no _ (underscore). Why no underscore? Because of
experience with typosquatting and human difficulties remembering which
is which in a package name. If we allow only one, we offer less room
for confusion. And of the two, I think - is prettier.

SO THERE