I like feeling validated. (#54)
* Validate legacy package names before attempting to fetch them.

Also, bumped deps. Removed some unused imports.

* A little bit of cli documentation

Also some pointers to things that need to be implemented.

* Groundwork for package & user name validations.

Wrap up the legacy package name validity check to hide the parts of that
API we don't care about. Sprinkle a little joi here & there.

More importantly, implement a hook for defining entropic's package name rules.

I have commenced the bikeshedding with an opinion: a-z, 0-0, and -.
Controversially, no _ (underscore). Why no underscore? Because of
experience with typosquatting and human difficulties remembering which
is which in a package name. If we allow only one, we offer less room
for confusion. And of the two, I think - is prettier.

SO THERE

* Package name validation needs to consider the legacy namespace.

Made it do so.

While I was in a persnickety mood, wrote some tests for it.