WIP: Invitations & joining. (#72)
* Invitations & joining, plus a consistency sweep.

Invitations are a boolean on maintainers, the `accepted` field. We now
set that to true on package creation for the creator. Creating an
invitation makes a maintainership with active false & accepted false.
On accept, both fields flip to true. On decline, the maintainership is
removed. (I see an abuse hole here for repeated invitations, but we can
close it later.)

Pulled out some re-usable decorators from the package maintainers file.
These will come in handy when I implement inviting to namespaces.

Smacked the umbrella of consistency against our server->client messaging.
We now use the `message` field and wrap all text in an object. Lists of
things are now in an `objects` field. Single-object responses remain as
they are.

Switched loggers from pino to bole so I can namespace logs.

Cleaned up our lint process a bit by pulling both the unused .eslintrc
and the in-use .prettierrc to the top level of the repo.

* Merrrrrrrrging nothing more than merging

* These checks are redundant with the decorators.

* Clean up some error/message field stuff.

* Implemented maintainership lists as well as invitation lists.

* host host host host host host host