The cheese moves.
<ricard strauss organ pedal note>

<trumpet blast>

Laying the groundwork here for multiple services in one. I think we
are going to cause problems for our users if we keep trying to do
everything in one process, and it'll be easier overall if we break
things up by responsibility now.

Better to get the chaos over with now.