A proposed workflow for Entropic.
This is a discussion-based process for analyzing problems and
making decisions about how to solve them. It's not intended to
be followed dogmatically, but instead provide guidelines for how
to approach making a proposal to the project.

It's my attempt to codify the process Chris & I loosely follow in
most of the work we've done together, including the behind the scenes
work we've done on this project. Now that we are doing the work with
more people, including lots of people we haven't worked with before,
it seems important to me to write it down.