feat: IPv6 support
Since I'm on a weird network where loopback routing doesn't work because
NAT is frustrating, but IPv6 gives me a global namespace, my copy of
entropic registry is running on a host that is only consistently
available via IPv6.

The defaults for the http and https agents in node are frustratingly not
IPv6 inclusive, but `node-fetch` does let us thread through our own
configuration.

I opted not to change the global agent, but instead wrap fetch. I could
see the other option working as well and perhaps less invasively, but
would be completely process global and would affect more than registry
connections if any were to be made.