cli/lib/main now rejects command names with file metachars
Fixes #251

This adds a check to function main so that when the first argument,
which is interpreted as a command name, has a file separator, it is
rejected, logged, and the command `help` is used instead.

To ease testing, this splits the analysis code out of `function main`
into `function unpack` which find the `cmd` to invoke but leaves the
invoking and error recovery to `main`.

This adds tests that a dodgy command is rejected, but that others like
`whoami` continue to function.

```sh
$ (pushd cli; npm t; popd)
$ npm run lint
```